minor changes to DB code
